Add files via upload
This repository contains my personal portfolio website showcasing my academic background, projects, and technical skills. The site is built using HTML, CSS, and JavaScript with a fully responsive design for both desktop and mobile. It highlights my MCA journey, certifications, and hands-on projects in web development, Python, Django, and machine learning. The portfolio is designed to present my work in a professional way and serve as a platform for recruiters, collaborators, and potential employers to connect with me